DS_core_20250325_062717/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
numpy>=1.24.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
beautifulsoup4>=4.10.0
aiohttp>=3.8.1 